Movie_R1/requirements.txt
alembic==1.10.2
altair==4.2.2
attrs==22.2.0
beautifulsoup4==4.12.2
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.7
decorator==5.1.1
docker==6.1.2
entrypoints==0.4
Faker==18.9.0
favicon==0.7.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
hashids==1.3.1
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
mlflow==2.3.2
more-itertools==9.1.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
protobuf==3.20.3
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.7.0
pymdown-extensions==10.0.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.7
sqlparse==0.4.4
st-annotated-text==4.0.0
streamlit==1.19.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tabulate==0.9.0
tenacity==8.2.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
waitress==2.1.2
watchdog==2.3.0
websocket-client==1.5.1
Werkzeug==2.2.3
zipp==3.14.0
